match target data type to be float

+++ fairseq_signals/models/clocs/clocs_dc.py
@@ -97,7 +97,7 @@ class ClocsDcModel(BaseModel):
 
         target = target.expand(x.size()).contiguous().view(-1, x.size(-1))
 
-        return target
+        return target.float()
     
     def get_normalized_probs(self, net_output, log_probs):
         """Get normalized probabilities (or log probs) from a net's output."""
